move echox to httpserve/middleware
