--- internal/echox/doc.go
@@ -1,2 +1,0 @@
-// Package echox adds helper functions for the echo server
-package echox
