+++ internal/httpserve/middleware/echomw/doc.go
@@ -1,0 +1,2 @@
+// Package echomw adds middelware specific to echo
+package echomw
