+++ internal/httpserve/middleware/echomw/echo.go
@@ -1,4 +1,4 @@
-package echox
+package echomw
 
 import (
 	"context"
