+++ internal/httpserve/middleware/echomw/test_tools.go
@@ -1,4 +1,4 @@
-package echox
+package echomw
 
 import (
 	"net/http"
